Current setup works with import statements
