+++ .eslintrc.yml
@@ -3,6 +3,8 @@ env:
   browser: true
   es6: true
   jquery: true
+parserOptions:
+  sourceType: module
 rules:
   no-console:
     - off
