+++ src/main/resources/static/javascript/widget-components.jsx
